feat: Check position dataframe (assertions)
Checks if dataframe contains a specific name on a certain position. This is done for the Seahorse "Assay Information" sheet, to make sure that every required information to perform the analysis is available.

For example: we will check if the name "Assay Name" is given for postion 4 of the dataframe derived from the "Assay information" sheet.

Note:
R/utils.R and assertion/assertions_read.R contain simular functions regarding this commit. The functions in R/utils are meant to be used in a more general way. The functions in assertions/assertions_read.R are specific to the meta_df derived from the "Assay information" sheet. This to make clear that the functions belong to assertions.