refactor: Move validations to R directory.
When the scripts for validation are in a /assertion branch, the .R functions won't be exported to the NAMESPACE file and the functions will not be recognized. However, when the validate scripts are in the /R directory, the functions will be exported to the NAMESPACE file which causes these functions to be recognised by other functions.